requirements.txt
beautifulsoup4==4.7.1
Django==2.1.7
django-ckeditor==5.6.1
django-filter==2.1.0
django-htmlmin==0.10.0
django-js-asset==1.2.1
django-robots==3.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
gunicorn==19.9.0
html5lib==1.0.1
Khayyam==3.0.17
Markdown==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
six==1.12.0
soupsieve==1.8
webencodings==0.5.1
